+++ docs/docs/examples/rag/embeddings.md
@@ -9,7 +9,7 @@ sidebar_position: 40
 In order to store the data in our vector database, we need to convert the free text into embeddings. Let's look at GitHub first. First when we are extracting data from GitHub, we will want to be careful to avoid rate limiting. We can do this by partitioning our asset into weekly chunks using a `WeeklyPartitionsDefinition`:
 
 <CodeExample
-  path="docs_projects/project_ask_ai_dagster/project_ask_ai_dagster/assets/ingestion.py"
+  path="docs_projects/project_ask_ai_dagster/project_ask_ai_dagster/defs/assets/ingestion.py"
   language="python"
   startAfter="start_partition"
   endBefore="end_partition"
@@ -18,7 +18,7 @@ In order to store the data in our vector database, we need to convert the free t
 We will supply that partition in the decorator of our asset, as well as an <PyObject section="assets" module="dagster" object="AutomationCondition" /> to ensure that the weekly partition is updated every Monday. The rest of our asset will use the `GithubResource` we defined earlier and return a list of LangChain `Documents`:
 
 <CodeExample
-  path="docs_projects/project_ask_ai_dagster/project_ask_ai_dagster/assets/ingestion.py"
+  path="docs_projects/project_ask_ai_dagster/project_ask_ai_dagster/defs/assets/ingestion.py"
   language="python"
   startAfter="start_github_issues_raw"
   endBefore="end_github_issues_raw"
@@ -27,7 +27,7 @@ We will supply that partition in the decorator of our asset, as well as an <PyOb
 The next asset will convert those `Documents` to vectors and upload them to Pinecone. In order to generate the embeddings, we will need an AI model. In this case, we will use OpenAI's `text-embedding-3-small` model to transform the text we have collected from GitHub into embeddings. Dagster provides an `OpenAIResource` to interact with the OpenAI client ,and we will use that to create the embeddings. This asset will also create the index in Pinecone:
 
 <CodeExample
-  path="docs_projects/project_ask_ai_dagster/project_ask_ai_dagster/assets/ingestion.py"
+  path="docs_projects/project_ask_ai_dagster/project_ask_ai_dagster/defs/assets/ingestion.py"
   language="python"
   startAfter="start_github_issues_embeddings"
   endBefore="end_github_issues_embeddings"
@@ -40,7 +40,7 @@ This process will be very similar for the GitHub discussions content.
 Looking at the code, you may have noticed the `document_io_manager`. Because LangChain `Documents` are a special object type, we need to do some work to serialize and deserialize the data. [I/O Managers](/guides/build/io-managers/) are responsible for handling the inputs and outputs of assets and how the data is persisted. This I/O manager will use the local file system to save the output of assets returning `Documents` as JSON files. It will then read those JSON files back into `Documents` in assets that take in those inputs:
 
 <CodeExample
-  path="docs_projects/project_ask_ai_dagster/project_ask_ai_dagster/definitions.py"
+  path="docs_projects/project_ask_ai_dagster/project_ask_ai_dagster/defs/io_managers.py"
   language="python"
   startAfter="start_io_manager"
   endBefore="end_io_manager"
@@ -62,7 +62,7 @@ The assets for the documentation scraping will behave similar to the GitHub asse
 The asset that generates the embeddings with the documentation site will need one additional change. Because the content of the documentation pages is so large, we need to split data into chunks. The `split_text` function ensures that we split the text into equal length chunks. We also want to keep similar chunks together and associated with the page they were on so we will hash the index of the URL to ensure data stays together. correctly Once the data is chunked, it can be batched and sent to Pinecone:
 
 <CodeExample
-  path="docs_projects/project_ask_ai_dagster/project_ask_ai_dagster/assets/ingestion.py"
+  path="docs_projects/project_ask_ai_dagster/project_ask_ai_dagster/defs/assets/ingestion.py"
   language="python"
   startAfter="start_batch"
   endBefore="end_batch"
