+++ examples/docs_projects/project_ask_ai_dagster/project_ask_ai_dagster/defs/assets/retrieval.py
@@ -1,12 +1,12 @@
 import dagster as dg
 from dagster_openai import OpenAIResource
 
-from project_ask_ai_dagster.assets.ingestion import (
+from project_ask_ai_dagster.defs.assets.ingestion import (
     docs_embedding,
     github_discussions_embeddings,
     github_issues_embeddings,
 )
-from project_ask_ai_dagster.resources.pinecone import PineconeResource
+from project_ask_ai_dagster.defs.resources.pinecone import PineconeResource
 
 
 # start_config
