+++ examples/docs_projects/project_ask_ai_dagster/project_ask_ai_dagster/defs/resources/pinecone.py
