+++ examples/docs_projects/project_ask_ai_dagster/project_ask_ai_dagster_tests/test_defs.py
@@ -1,0 +1,5 @@
+from project_ask_ai.definitions import defs
+
+
+def test_def():
+    assert defs
