--- examples/docs_projects/project_ask_ai_dagster/requirements.txt
@@ -1,12 +1,0 @@
-langchain
-langchain-core
-gql
-langchain-community
-langchain-openai
-dagster
-dagster-webserver
-tenacity
-bs4
-lxml
-dagster-openai
-pinecone
