+++ examples/docs_projects/project_ask_ai_dagster/setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+name = project_ask_ai_dagster
