+++ examples/docs_projects/project_ask_ai_dagster/setup.py
@@ -1,0 +1,14 @@
+from setuptools import find_packages, setup
+
+setup(
+    name="project_llm_fine_tune",
+    packages=find_packages(exclude=["project_llm_fine_tune_tests"]),
+    install_requires=[
+        "dagster",
+        "dagster-openai",
+        "dagster-duckdb",
+        "pandas",
+        "openai",
+    ],
+    extras_require={"dev": ["dagster-webserver", "pytest", "ruff"]},
+)
